added separate Header component
